Moving early-init.el to init.el
Matching pattern laid out by elpaca. One benefit is the GUI is up
nearly instantly while elpaca is doing its thing.

The other thing added here is the usage of `locate-user-emacs-file`
which will now make this just work if we want to test out our config
in a different directory.